Add Demote, Unban, and Cleanup commands
Added a demote command for owners. Owners can now change the rank of a mod or admin back to the regular user rank.

Added an unban command. Owners can now unban users by inputing the user's name. This makes a simpler alternative for unbanning users, instead of having to do it manually through the database.
(May rewrite the getBlacklistByName function at a later time, but it works currently)

Incorporated a cleanup command (acquired from somebody else; I do not know who wrote it). In order for the commmand to work, it must be ran after blacklisting a user in the chat. The bot will then make a list of all the MSIDs from that user and delete them.