Add /preban, reply-to for /say commands, reasons for /delete and /warn
Added a preemptive blacklist command for owners. Owners can type /preban [USERID]:[REASON] to preemptively ban users already in the database, without needed to blacklist their post in the bot.
This command can also blacklist users who have not used the bot by adding their Telegram ID to the database and setting their rank to -10 (this will not send a ERR_BLACKLISTED reply to the user not in the database)
A [REASON] must be given in order for the command to work.

Added reply-to for the ranked /say commands. Admins, mods, owners, and sysop have the option of replying to messages with the official signed message.

Added reasons for /delete and warn commands. Mods can give a reason for cooldowning/deleting posts if they choose (similar to the /remove command).

Added a getUserByID function specifically for the /preban command
Reorderd cmds[] in telegram.py, just to look neater.
Made reply MESSAGE_REMOVED less verbose.
Changed /help commands to reflect the new additions
Removed getBlacklistedByName function and just incorporated it into getUserByName