Added voice message spam filter and a "count down" for the media limit
The voice message spam filter begins counting on the first message. The user can send up to 10 messages within the span of half an hour. If the user exceeds 10 messages then, he will be unable to send any more voice messages until the half hour is up.
The spam filter can be toggled through the config file.

The media limit calculates the amount of time left before a new user can begin sending messages, and sends the result to him.  This will say something along the lines of, "Try again 4.8 hours later"
This is probably an odd way to show the amount of time left for the media limit, though.